Safari-609.1.11.1/LayoutTests/svg/dynamic-updates/SVGFESpotLightElement-svgdom-x-prop-expected.txt

SVG 1.1 dynamic update tests

Tests dynamic updates of the 'x' property of the SVGFESpotLightElement object

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS spotLightElement.x.baseVal is 0
PASS spotLightElement.x.baseVal is 113
PASS successfullyParsed is true

TEST COMPLETE
